TST: follow-up to Test nested pandas array #24993 (#25155)
* revert changes to tests in gh-24993

* Test nested PandasArray

* isort test_numpy.py

* change NP_VERSION_INFO

* use LooseVersion

* add _np_version_under1p16

* remove blank line from merge master

* add doctstrings to fixtures
