+++ pandas/tests/extension/base/missing.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pytest
 
 import pandas as pd
 import pandas.util.testing as tm
@@ -89,14 +88,13 @@ class BaseMissingTests(BaseExtensionTests):
         result = ser.fillna(ser)
         self.assert_series_equal(result, ser)
 
-    @pytest.mark.parametrize('method', ['ffill', 'bfill'])
-    def test_fillna_series_method(self, data_missing, method):
+    def test_fillna_series_method(self, data_missing, fillna_method):
         fill_value = data_missing[1]
 
-        if method == 'ffill':
+        if fillna_method == 'ffill':
             data_missing = data_missing[::-1]
 
-        result = pd.Series(data_missing).fillna(method=method)
+        result = pd.Series(data_missing).fillna(method=fillna_method)
         expected = pd.Series(data_missing._from_sequence(
             [fill_value, fill_value], dtype=data_missing.dtype))
