+++ pandas/tests/extension/conftest.py
@@ -2,6 +2,8 @@ import operator
 
 import pytest
 
+from pandas import Series
+
 
 @pytest.fixture
 def dtype():
@@ -107,4 +109,59 @@ def data_for_grouping():
 @pytest.fixture(params=[True, False])
 def box_in_series(request):
     """Whether to box the data in a Series"""
+    return request.param
+
+
+@pytest.fixture(params=[
+    lambda x: 1,
+    lambda x: [1] * len(x),
+    lambda x: Series([1] * len(x)),
+    lambda x: x,
+], ids=['scalar', 'list', 'series', 'object'])
+def groupby_apply_op(request):
+    """
+    Functions to test groupby.apply().
+    """
+    return request.param
+
+
+@pytest.fixture(params=[True, False])
+def as_frame(request):
+    """
+    Boolean fixture to support Series and Series.to_frame() comparison testing.
+    """
+    return request.param
+
+
+@pytest.fixture(params=[True, False])
+def as_series(request):
+    """
+    Boolean fixture to support arr and Series(arr) comparison testing.
+    """
+    return request.param
+
+
+@pytest.fixture(params=[True, False])
+def use_numpy(request):
+    """
+    Boolean fixture to support comparison testing of ExtensionDtype array
+    and numpy array.
+    """
+    return request.param
+
+
+@pytest.fixture(params=['ffill', 'bfill'])
+def fillna_method(request):
+    """
+    Parametrized fixture giving method parameters 'ffill' and 'bfill' for
+    Series.fillna(method=<method>) testing.
+    """
+    return request.param
+
+
+@pytest.fixture(params=[True, False])
+def as_array(request):
+    """
+    Boolean fixture to support ExtensionDtype _from_sequence method testing.
+    """
     return request.param
