+++ pandas/tests/extension/test_sparse.py
@@ -287,11 +287,10 @@ class TestMethods(BaseSparseTests, base.BaseMethodsTests):
             pytest.skip("TODO(SparseArray.__setitem__ will preserve dtype.")
         super(TestMethods, self).test_combine_first(data)
 
-    @pytest.mark.parametrize("as_series", [True, False])
     def test_searchsorted(self, data_for_sorting, as_series):
         with tm.assert_produces_warning(PerformanceWarning):
             super(TestMethods, self).test_searchsorted(data_for_sorting,
-                                                       as_series=as_series)
+                                                       as_series)
 
 
 class TestCasting(BaseSparseTests, base.BaseCastingTests):
